+++ src/main/java/de/alpharogroup/collections/map/MapFactory.java
@@ -24,13 +24,10 @@
  */
 package de.alpharogroup.collections.map;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.TreeMap;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -40,13 +37,13 @@ import org.apache.commons.collections4.map.LazyMap;
 import lombok.experimental.UtilityClass;
 
 /**
- * Extensions class for use with Map objects.
+ * Factory class for create new Map objects
  *
  * @version 1.0
  * @author Asterios Raptis
  */
 @UtilityClass
-public final class MapExtensions
+public final class MapFactory
 {
 
 	/**
@@ -57,68 +54,11 @@ public final class MapExtensions
 	 * @param <V>
 	 *            the generic type of the value
 	 *
-	 * @return The new {@link InsertionOrderMap}
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
+	 * @return The new {@link InsertionOrderMap}.
 	 */
 	public static <K, V> Map<K, V> newInsertionOrderMap()
 	{
 		return new InsertionOrderMap<>();
-	}
-
-	/**
-	 * Returns the first founded key from the given value or null if nothing is found.
-	 *
-	 * @param <K>
-	 *            the generic type of the key
-	 * @param <V>
-	 *            the generic type of the value
-	 * @param map
-	 *            The Map.
-	 * @param value
-	 *            The value.
-	 * @return Returns the first key from the given value or null if nothing found.
-	 */
-	public static <K, V> K getKeyFromValue(final Map<K, V> map, final V value)
-	{
-		K key = null;
-		for (final Entry<K, V> entry : map.entrySet())
-		{
-			key = entry.getKey();
-			if (map.get(key).equals(value))
-			{
-				break;
-			}
-		}
-		return key;
-	}
-
-	/**
-	 * Returns a Collection from all founded keys from the given value or null if nothing found.
-	 *
-	 * @param <K>
-	 *            the generic type of the key
-	 * @param <V>
-	 *            the generic type of the value
-	 * @param map
-	 *            The Map.
-	 * @param value
-	 *            The value.
-	 * @return Returns the key from the given value or an empty Collection if nothing found.
-	 */
-	public static <K, V> Collection<K> getKeysFromValue(final Map<K, V> map, final V value)
-	{
-		final Collection<K> keys = new ArrayList<>();
-		for (final Entry<K, V> entry : map.entrySet())
-		{
-			final V val = entry.getValue();
-			if (val.equals(value))
-			{
-				final K key = entry.getKey();
-				keys.add(key);
-			}
-		}
-		return keys;
 	}
 
 	/**
@@ -147,8 +87,6 @@ public final class MapExtensions
 	 * </pre>
 	 *
 	 * @return The {@link java.util.Map} that acts like a javascript associative array.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static Map<Integer, Map<String, String>> newAssosiativeArrayMap()
 	{
@@ -164,8 +102,6 @@ public final class MapExtensions
 	 *            the generic type of the value
 	 *
 	 * @return The new {@link ConcurrentHashMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static <K, V> ConcurrentHashMap<K, V> newConcurrentHashMap()
 	{
@@ -182,8 +118,6 @@ public final class MapExtensions
 	 * @param initialCapacity
 	 *            the initial capacity
 	 * @return The new {@link ConcurrentHashMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static <K, V> ConcurrentHashMap<K, V> newConcurrentHashMap(final int initialCapacity)
 	{
@@ -199,8 +133,6 @@ public final class MapExtensions
 	 *            the generic type of the value
 	 *
 	 * @return The new {@link HashMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static <K, V> Map<K, V> newHashMap()
 	{
@@ -216,8 +148,6 @@ public final class MapExtensions
 	 *            the generic type of the value
 	 *
 	 * @return The new {@link LinkedHashMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static <K, V> Map<K, V> newLinkedHashMap()
 	{
@@ -235,8 +165,6 @@ public final class MapExtensions
 	 *            the initial capacity
 	 *
 	 * @return The new {@link LinkedHashMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static <K, V> Map<K, V> newLinkedHashMap(final int initialCapacity)
 	{
@@ -253,8 +181,6 @@ public final class MapExtensions
 	 * @param initialCapacity
 	 *            the initial capacity
 	 * @return The new {@link HashMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static <K, V> Map<K, V> newHashMap(final int initialCapacity)
 	{
@@ -271,8 +197,6 @@ public final class MapExtensions
 	 *            the generic type of the value
 	 *
 	 * @return The new {@link LazyMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	@SuppressWarnings({ "unchecked", "rawtypes" })
 	public static <K, V> Map<K, V> newLazyMap()
@@ -290,8 +214,6 @@ public final class MapExtensions
 	 *            the generic type of the value
 	 *
 	 * @return The new {@link LazyMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	@SuppressWarnings({ "unchecked", "rawtypes" })
 	public static <K, V> Map<K, V> newLazyTreeMap()
@@ -311,8 +233,6 @@ public final class MapExtensions
 	 *            the comparator
 	 *
 	 * @return The new {@link LazyMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	@SuppressWarnings({ "unchecked", "rawtypes" })
 	public static <K, V> Map<K, V> newLazyTreeMap(final Comparator<? super K> comparator)
@@ -330,8 +250,6 @@ public final class MapExtensions
 	 *            the generic type of the value
 	 *
 	 * @return The new {@link TreeMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static <K, V> Map<K, V> newTreeMap()
 	{
@@ -348,45 +266,10 @@ public final class MapExtensions
 	 * @param comparator
 	 *            the comparator
 	 * @return The new {@link TreeMap}.
-	 * @deprecated use instead the same name method from <code>MapFactory</code>
-	 * Note: will be removed with the next minor release
 	 */
 	public static <K, V> Map<K, V> newTreeMap(final Comparator<? super K> comparator)
 	{
 		return new TreeMap<>(comparator);
-	}
-
-	/**
-	 * Converts a two dimensional Array to a Map.
-	 *
-	 * @param <T>
-	 *            the generic type
-	 * @param twoDimArray
-	 *            The two dimensional Array.
-	 * @return The map produced from the two dimensional Array.
-	 */
-	public static <T> Map<T, T> toGenericMap(final T[][] twoDimArray)
-	{
-		final Map<T, T> map = new LinkedHashMap<>();
-		for (final T[] element : twoDimArray)
-		{
-			final T key = element[0];
-			final T value = element[1];
-			map.put(key, value);
-		}
-		return map;
-	}
-
-	/**
-	 * Converts a two dimensional String Array to a Map.
-	 *
-	 * @param twoDimArray
-	 *            The two dimensional String Array.
-	 * @return The map produced from the two dimensional String Array.
-	 */
-	public static Map<String, String> toMap(final String[][] twoDimArray)
-	{
-		return toGenericMap(twoDimArray);
 	}
 
 }