+++ src/test/java/de/alpharogroup/collections/list/ListExtensionsTest.java
@@ -337,6 +337,7 @@ public class ListExtensionsTest
 	/**
 	 * Test the method {@link ListExtensions#newRangeArray(int, int)}
 	 */
+	@SuppressWarnings("deprecation")
 	@Test
 	public void testNewRangeArray()
 	{
@@ -358,6 +359,7 @@ public class ListExtensionsTest
 	 * Test the method {@link ListExtensions#newRangeArray(int, int)} where end is smaller then
 	 * start
 	 */
+	@SuppressWarnings("deprecation")
 	@Test(expectedExceptions = IllegalArgumentException.class)
 	public void testNewRangeArrayException()
 	{